+++ ios/MullvadVPN/View controllers/SelectLocation/DataSource/LocationNode.swift
@@ -22,6 +22,7 @@ class LocationNode: @unchecked Sendable {
     var isConnected: Bool
     var isSelected: Bool
     var isExcluded: Bool
+    let id = UUID()
 
     init(
         name: String,
@@ -48,6 +49,28 @@ class LocationNode: @unchecked Sendable {
         self.isSelected = isSelected
         self.isExcluded = isExcluded
     }
+
+    /// Recursively copies a node, its parent and its descendants from another
+    /// node (tree), with an optional custom root parent.
+    func copy(withParent parent: LocationNode? = nil) -> LocationNode {
+        let node = LocationNode(
+            name: name,
+            code: code,
+            locations: locations,
+            isActive: isActive,
+            parent: parent,
+            children: [],
+            showsChildren: showsChildren,
+            isHiddenFromSearch: isHiddenFromSearch,
+            isConnected: isConnected,
+            isSelected: false,  // explicity set to false since it's a different node
+            isExcluded: isExcluded
+        )
+
+        node.children = recursivelyCopyChildren(withParent: node)
+
+        return node
+    }
 }
 
 extension LocationNode {
@@ -55,6 +78,9 @@ extension LocationNode {
         parent?.root ?? self
     }
 
+    var asRecentLocationNode: RecentLocationNode? {
+        self as? RecentLocationNode
+    }
     var asCustomListNode: CustomListLocationNode? {
         self as? CustomListLocationNode
     }
@@ -126,36 +152,19 @@ extension LocationNode {
                 }
             }
     }
-}
 
-extension LocationNode {
-    /// Recursively copies a node, its parent and its descendants from another
-    /// node (tree), with an optional custom root parent.
-    func copy(withParent parent: LocationNode? = nil) -> LocationNode {
-        let node = LocationNode(
-            name: name,
-            code: code,
-            locations: locations,
-            isActive: isActive,
-            parent: parent,
-            children: [],
-            showsChildren: showsChildren,
-            isHiddenFromSearch: isHiddenFromSearch,
-            isConnected: isConnected,
-            isSelected: false,  // explicity set to false since it's a different node
-            isExcluded: isExcluded
-        )
-
-        node.children = recursivelyCopyChildren(withParent: node)
-
-        return node
+    func pathToRoot() -> [String] {
+        var path: [String] = [name]
+        forEachAncestor { locationNode in
+            path.insert(NSLocalizedString(locationNode.name, comment: ""), at: 0)
+        }
+        return path
     }
 
-    private func recursivelyCopyChildren(withParent parent: LocationNode) -> [LocationNode] {
+    fileprivate func recursivelyCopyChildren(withParent parent: LocationNode) -> [LocationNode] {
         children.map { $0.copy(withParent: parent) }
     }
 }
-
 extension LocationNode: Hashable {
     func hash(into hasher: inout Hasher) {
         hasher.combine(code)
@@ -203,6 +212,56 @@ class CustomListLocationNode: LocationNode, @unchecked Sendable {
         customList: CustomList
     ) {
         self.customList = customList
+
+        super.init(
+            name: name,
+            code: code,
+            locations: locations,
+            isActive: isActive,
+            parent: parent,
+            children: children,
+            showsChildren: showsChildren,
+            isHiddenFromSearch: isHiddenFromSearch
+        )
+    }
+
+    /// Recursively copies a node, its parent and its descendants from another
+    /// node (tree), with an optional custom root parent.
+    override func copy(withParent parent: LocationNode? = nil) -> LocationNode {
+        let node = CustomListLocationNode(
+            name: name,
+            code: code,
+            locations: locations,
+            isActive: isActive,
+            parent: parent,
+            children: [],
+            showsChildren: showsChildren,
+            isHiddenFromSearch: isHiddenFromSearch,
+            customList: customList
+        )
+
+        node.children = recursivelyCopyChildren(withParent: node)
+
+        return node
+    }
+
+}
+
+class RecentLocationNode: LocationNode, @unchecked Sendable {
+    let locationInfo: [String]?
+
+    init(
+        name: String,
+        code: String,
+        locations: [RelayLocation] = [],
+        isActive: Bool = true,
+        parent: LocationNode? = nil,
+        children: [LocationNode] = [],
+        showsChildren: Bool = false,
+        isHiddenFromSearch: Bool = false,
+        locationInfo: [String]?
+    ) {
+        self.locationInfo = locationInfo
 
         super.init(
             name: name,